requirements.txt
# Venue Booking System - Python Dependencies

# Web Framework
Flask==3.0.0

# Supabase Client
supabase==2.3.0
postgrest-py==0.13.2

# Environment Variables
python-dotenv==1.0.0

# Date and Time Handling
python-dateutil==2.8.2

# HTTP Requests
requests==2.31.0

# Authentication & Security
PyJWT==2.8.0
cryptography==41.0.7

# Database (PostgreSQL client, optional for direct access)
psycopg2-binary==2.9.9

# Production Server
gunicorn==21.2.0